@@ -32,10 +32,13 @@
     "@radix-ui/react-toast": "^1.1.4",
     "@radix-ui/react-toggle": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.6",
+    "@t3-oss/env-core": "^0.6.1",
     "@t3-oss/env-nextjs": "^0.3.1",
     "@tanstack/react-query": "^4.29.12",
     "@types/node": "20.2.3",
-    "@types/react-dom": "18.2.4",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.4",
+    "@vercel/edge-config": "^0.2.1",
     "@visx/axis": "^3.3.0",
     "@visx/event": "^3.3.0",
     "@visx/group": "^3.3.0",
@@ -43,6 +46,7 @@
     "@visx/scale": "^3.3.0",
     "@visx/shape": "^3.3.0",
     "@visx/tooltip": "^3.3.0",
+    "@visx/voronoi": "^3.3.0",
     "autoprefixer": "10.4.14",
     "class-variance-authority": "^0.6.0",
     "clsx": "^1.2.1",
@@ -51,20 +55,22 @@
     "echarts": "^5.4.2",
     "fast-deep-equal": "^3.1.3",
     "fuse.js": "^6.6.2",
+    "html-to-image": "^1.11.11",
     "lucide-react": "^0.258.0",
-    "next": "^13.4.12",
+    "next": "^13.4.16",
     "next-themes": "^0.2.1",
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-day-picker": "^8.7.1",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.44.2",
+    "react-inlinesvg": "^4.0.2",
     "react-xarrows": "^2.0.2",
     "tailwind-merge": "^1.12.0",
     "tailwindcss": "3.3.2",
     "tailwindcss-animate": "^1.0.5",
-    "typescript": "5.0.4",
-    "zod": "^3.21.4"
+    "typescript": "^5.1.6",
+    "zod": "^3.22.1"
   },
   "devDependencies": {
     "@storybook/addon-essentials": "^7.0.20",
@@ -76,6 +82,7 @@
     "@storybook/react": "^7.0.20",
     "@storybook/testing-library": "^0.0.14-next.2",
     "@svgr/webpack": "^8.0.1",
+    "@tanstack/eslint-plugin-query": "^4.32.5",
     "eslint": "8.43.0",
     "eslint-config-next": "13.4.6",
     "eslint-plugin-storybook": "^0.6.13",
