@@ -2,6 +2,7 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ThemeProvider } from "next-themes";
 import { useState } from "react";
+import { TooltipProvider } from "./ui/Tooltip";
 
 type RootProps = {
   children: React.ReactNode;
@@ -17,7 +18,11 @@ export default function RQProvider({ children }: RootProps) {
   );
   return (
     <ThemeProvider attribute="class">
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+      <TooltipProvider delayDuration={300}>
+        <QueryClientProvider client={queryClient}>
+          {children}
+        </QueryClientProvider>
+      </TooltipProvider>
     </ThemeProvider>
   );
 }
