@@ -1,18 +1,18 @@
 import { useEffect, useState } from "react";
 
-/**
- * WARN: TESTS NEEDED
- */
-export function useLocalStorage<T>(key: string, fallbackValue: T | undefined) {
+export function useLocalStorage<T>(key: string, fallbackValue?: T) {
   const [value, setValue] = useState(fallbackValue);
   useEffect(() => {
     const stored = localStorage.getItem(key);
-    if (stored !== "undefined")
+    if (stored !== "undefined") {
       setValue(stored ? JSON.parse(stored) : fallbackValue);
-  }, [fallbackValue, key]);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   useEffect(() => {
     localStorage.setItem(key, JSON.stringify(value));
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [key, value]);
 
   return [value, setValue] as const;
