@@ -1,7 +1,7 @@
 const STORAGE = {
   jadeEstimateForm: "jadeEstimateForm",
-  gachaForm: 'gachaForm',
-  uid: "uid"
-} as const
-export type Storage = (typeof STORAGE)[keyof typeof STORAGE]
-export default STORAGE
+  gachaForm: "gachaForm",
+  playerProfiles: "playerProfiles",
+} as const;
+export type Storage = (typeof STORAGE)[keyof typeof STORAGE];
+export default STORAGE;
